🚚 Rename : 결제 관련 페이지 하나의 폴더로 합치기 #42
