+++ src/routes/Router.tsx
@@ -20,14 +20,14 @@ import ReservationPage from '@pages/ReservationPage';
 import HostNotiPage from '@pages/HostNotiPage';
 import UserInfoPage from '@pages/UserInfoPage';
 import PaymentPage from '@pages/PaymentPage';
-import PaymentSuccessPage from '@pages/PaymentSuccessPage';
+import PaymentSuccessPage from '@pages/PaymentPage/PaymentSuccessPage';
 import HostInfoPage from '@pages/HostInfoPage';
 import HostInfoEditPage from '@pages/HostInfoEditPage';
 import UserInfoEditPage from '@pages/UserInfoEditPage';
 import DetailPage from '@pages/DetailPage';
 import ChatListPage from '@pages/ChatListPage';
 import ChatPage from '@pages/ChatPage';
-import PaymentFailPage from '@pages/PaymentFailPage';
+import PaymentFailPage from '@pages/PaymentPage/PaymentFailPage';
 
 const router = createBrowserRouter([
   {
